test: run the Tier-1 exactness suite in float64 (fix flaky O(3) invariance)
test_exact.py::test_energy_global_o3_invariant failed locally (passed in CI)
because the exactness suite ran on the float32 "scouting" grid. A *generic*
O(3) rotation of the energy rounds to ~4e-4 in float32 — over the suite's
rtol=1e-4 — so pass/fail depended on the platform's XLA rounding, not on
correctness. (The 90deg rotation is a bit-exact component swap, so it never
tripped; the energy functional is genuinely O(3)-invariant: deviation ~1e-12
in float64.)

These tests assert identities that hold EXACTLY up to float rounding, so they
must run where that rounding is ~1e-12. Add tests/conftest.py to enable
jax_enable_x64 for the session and pin test_exact's grid to float64. x64 only
*enables* the capability — every other test still builds float32 grids and is
unaffected (verified: full suite 260 passed, 3 skipped, 0 regressions).

+++ tests/test_exact.py
@@ -17,7 +17,10 @@ from jax_solitons.models.faddeev import CP1Constraint, S2Constraint
 from jax_solitons.seeds import rational_map_hopfion
 from jax_solitons.topology import area_form_plaquette, hopf_charge
 
-GRID = BoxGrid(N=24, L=8.0)
+# float64: these are EXACT identities, only meaningful at full precision (see
+# conftest's x64 enable). In float32 a generic O(3) rotation rounds to ~4e-4 and
+# trips the tight tolerances below; in float64 the deviation is ~1e-12.
+GRID = BoxGrid(N=24, L=8.0, dtype=jnp.float64)
 
 
 def _random_smooth_n(grid: BoxGrid, seed: int, corr=1.0):
